Models/Assessment models/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
typing-extensions==4.8.0
langchain==0.0.350
langchain-huggingface==0.0.6
langchain-community==0.0.10
pdfplumber==0.10.3
python-docx==1.1.0